Seperate gql queries from api files
